Challenge Module 1
For the module 1 challenge, we were expected to use the kickstarter sheet data to develop two charts, Outcomes based on launch date and Play outcomes based on goals, which are attached to this repo. A detailed explanation of what the purpose of this data analysis is included, along with challenges and difficulties encountered, and a results section.