Update README.md
For the Module 1 Challenge, we were expected to use data from the kickstarter sheet and use that data to determine outcomes of plays based on different ranges of goal amounts and by theater launch dates. Included in this repository is the excel file that includes the kickstarter datasheet, and associated graphs for Module 1. A written report of the kickstarter analysis is included to summarize the project, how the analysis was done, challenges and difficulties encountered, and conclusions of the project.